Rename and remove cofig option
